--- __mocks__/fileMock.js
@@ -1,2 +1,0 @@
-// __mocks__/fileMock.js
-module.exports = 'test-file-stub';
