--- __mocks__/styleMock.js
@@ -1,2 +1,0 @@
-// __mocks__/styleMock.js
-module.exports = {};
