--- src/setupTests.ts
@@ -1,72 +1,0 @@
-// src/setupTests.ts
-import '@testing-library/jest-dom';
-
-// Global mocks
-
-// Mock for IntersectionObserver often needed for UI tests
-const mockIntersectionObserver = jest.fn();
-mockIntersectionObserver.mockReturnValue({
-  observe: jest.fn(),
-  unobserve: jest.fn(),
-  disconnect: jest.fn(),
-});
-window.IntersectionObserver = mockIntersectionObserver;
-
-// Global fetch mock
-global.fetch = jest.fn(() =>
-  Promise.resolve({
-    ok: true,
-    status: 200,
-    json: () => Promise.resolve({}), // Default empty JSON success
-    text: () => Promise.resolve(''), // Default empty text success
-    headers: new Headers(),
-    redirected: false,
-    type: 'basic',
-    url: '',
-    clone: jest.fn(function() { return this; }), // Return self for clone
-    body: null,
-    bodyUsed: false,
-    arrayBuffer: jest.fn().mockResolvedValue(new ArrayBuffer(0)),
-    blob: jest.fn().mockResolvedValue(new Blob([])), // Pass empty array to Blob constructor
-    formData: jest.fn().mockResolvedValue(new FormData()),
-  } as unknown as Response) // Type assertion for Response
-);
-
-// Mock for localStorage and sessionStorage (if used)
-const localStorageMock = (() => {
-  let store: { [key: string]: string } = {};
-  return {
-    getItem: (key: string) => store[key] || null,
-    setItem: (key: string, value: string) => {
-      store[key] = value.toString();
-    },
-    removeItem: (key: string) => {
-      delete store[key];
-    },
-    clear: () => {
-      store = {};
-    },
-    key: (index: number) => Object.keys(store)[index] || null,
-    get length() {
-      return Object.keys(store).length;
-    }
-  };
-})();
-Object.defineProperty(window, 'localStorage', { value: localStorageMock });
-Object.defineProperty(window, 'sessionStorage', { value: { ...localStorageMock } }); // Can use the same mock or a separate one
-
-
-// You might want to mock other browser APIs if they are used and not supported by JSDOM:
-// Object.defineProperty(window, 'matchMedia', {
-//   writable: true,
-//   value: jest.fn().mockImplementation(query => ({
-//     matches: false,
-//     media: query,
-//     onchange: null,
-//     addListener: jest.fn(), // deprecated
-//     removeListener: jest.fn(), // deprecated
-//     addEventListener: jest.fn(),
-//     removeEventListener: jest.fn(),
-//     dispatchEvent: jest.fn(),
-//   })),
-// });
